requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
boto3==1.34.101
    # via -r requirements.in
botocore==1.34.101
    # via
    #   -r requirements.in
    #   boto3
    #   s3transfer
certifi==2024.12.14
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via requests
crispy-bootstrap5==0.7
    # via -r requirements.in
cryptography==44.0.0
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-config-url==0.1.1
    # via -r requirements.in
dj-database-url==2.1.0
    # via -r requirements.in
django==5.0.4
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   dj-database-url
    #   django-allauth
    #   django-cors-headers
    #   django-filter
    #   django-storages
    #   django-tables2
    #   djangorestframework
    #   djangorestframework-simplejwt
django-allauth==0.51.0
    # via -r requirements.in
django-cors-headers==3.14.0
    # via -r requirements.in
django-crispy-forms==1.14.0
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-filter==24.3
    # via -r requirements.in
django-storages==1.13.2
    # via -r requirements.in
django-tables2==2.7.5
    # via -r requirements.in
django-widget-tweaks==1.5.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
gunicorn==22.0.0
    # via -r requirements.in
idna==3.10
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.2
    # via gunicorn
psycopg2-binary==2.9.10
    # via -r requirements.in
pycparser==2.22
    # via cffi
pyjwt[crypto]==2.10.1
    # via
    #   django-allauth
    #   djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.0.1
    # via -r requirements.in
python3-openid==3.2.0
    # via django-allauth
pytz==2024.2
    # via djangorestframework
requests==2.32.3
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via django-allauth
s3transfer==0.10.4
    # via boto3
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via django
typing-extensions==4.12.2
    # via dj-database-url
tzdata==2024.2
    # via django
urllib3==2.3.0
    # via
    #   botocore
    #   requests
